Move recusrive normalizers tests to a separate file
No functional changes in the code. Code was copy-pasted from
test_basic_version_store.py to test_recursive_normalizers.py
